Add fixture for getting graphql app
I didn't find a way to override the extensions passed to the graphql schema
hence had to write the extensions and most of the setup stuffs. So I had to
write a new router pointing to different path other than the original one.
Functionality will still same though.

I could prevent all this by writing a custom execution context and passing
it to the schema and just override the extensions without setting up a router
but just using the schema doesn't have the network call feel hence the reason
why I had to do all this.

If you have a better way do, well to send me a PR +1